OCR_with_gemini/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pypdf2
chromadb